refactor: delete sharedcomponents namespace
